+++ HISTORY.md
@@ -10,6 +10,8 @@
   * Document::saveXML()
   * DocumentFragment::appendXML()
   * Element::setIdAttribute() (stubbed out)
+* Allow a final string argument to Document::createElement() and
+  Document::createElementNS() for PHP compatibility.
 
 ## Dodo 0.2.0 (2021-07-26)
 * Update to IDLeDOM 0.7.2.